Consolidate semver libraries (#326)
Switch semver library use to `github.com/blang/semver/v4` to be
consistent across various uses. This is the semver library used by
`sigs.k8s.io/controller-runtime`, so it's consistent and reduces
the number of dependencies.

This also simplifies and eliminates cross dependencies between the API
packages and the operator runtime.

Signed-off-by: SuperQ <superq@gmail.com>